% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutch_colors.R
\name{hutch_colors}
\alias{hutch_colors}
\title{Fred Hutch Colors}
\usage{
hutch_colors(...)
}
\arguments{
\item{...}{Color names.}
}
\description{
The official colors of the Fred Hutch Cancer Center in Seattle, Washington.
}
\examples{
# All the colors!
hutch_colors()

# A Subset of colors
hutch_colors("Bright Blue", "Vivid Yellow")
}
