% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutch_palette.R
\name{hutch_palette}
\alias{hutch_palette}
\title{Returns function to interpolate a Fred Hutch color palette}
\usage{
hutch_palette(colors = names(hutch_colors()), reverse = FALSE, ...)
}
\arguments{
\item{colors}{A vector of official Fred Hutch color names.
See \code{hutchpal::hutch_colors()}.}

\item{reverse}{Should the palette be reversed?}

\item{...}{Additional arguments to pass to \code{grDevices::colorRampPalette()}.}
}
\description{
Returns function to interpolate a Fred Hutch color palette
}
