% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_hutch.R
\name{scale_color_hutch}
\alias{scale_color_hutch}
\title{Color scale constructor for Fred Hutch colors}
\usage{
scale_color_hutch(colors = names(hutch_colors()), reverse = FALSE, ...)
}
\arguments{
\item{colors}{A vector of official Fred Hutch color names.
See \code{hutchpal::hutch_colors()}.}

\item{reverse}{Should the palette be reversed?}

\item{...}{Additional arguments passed to \code{ggplot2::scale_color_gradientn()}.}
}
\description{
Color scale constructor for Fred Hutch colors
}
