% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_hutch.R
\name{scale_fill_hutch}
\alias{scale_fill_hutch}
\title{Fill scale constructor for Fred Hutch colors}
\usage{
scale_fill_hutch(colors = names(hutch_colors()), reverse = FALSE, ...)
}
\arguments{
\item{colors}{A vector of official Fred Hutch color names.
See \code{hutchpal::hutch_colors()}.}

\item{reverse}{Should the palette be reversed?}

\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_manual()}.}
}
\description{
Fill scale constructor for Fred Hutch colors
}
